Add market route validation
