+++ app/src/components/market/market_fund_container.tsx
@@ -5,8 +5,6 @@ import { useConnectedWeb3Context } from '../../hooks/connectedWeb3'
 import { useQuestion } from '../../hooks/useQuestion'
 import { useMarketMakerData } from '../../hooks/useMarketMakerData'
 import { FullLoading } from '../common/full_loading'
-import { MarketNotFound } from '../common/market_not_found'
-import { useCheckContractExists } from '../../hooks/useCheckContractExists'
 
 interface Props {
   marketMakerAddress: string
@@ -16,11 +14,7 @@ const MarketFundContainer = (props: Props) => {
   const context = useConnectedWeb3Context()
 
   const { marketMakerAddress } = props
-
-  const contractExists = useCheckContractExists(marketMakerAddress, context)
-
   const { question, resolution } = useQuestion(marketMakerAddress, context)
-
   const {
     totalPoolShares,
     userPoolShares,
@@ -30,10 +24,6 @@ const MarketFundContainer = (props: Props) => {
     marketMakerUserFunding,
     collateral,
   } = useMarketMakerData(marketMakerAddress, context)
-
-  if (!contractExists) {
-    return <MarketNotFound />
-  }
 
   if (!collateral) {
     return <FullLoading />
